requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
pillow==10.4.0
pyTelegramBotAPI==4.22.1
requests==2.32.3
telebot==0.0.5
urllib3==2.2.2
wget==3.2
